requirements.txt
django-compat==0.0.1
